+++ fltk-derive/src/widget.rs
@@ -67,6 +67,10 @@ pub fn impl_widget_base_trait(ast: &DeriveInput) -> TokenStream {
                 }
             }
 
+            fn default_fill() -> Self {
+                Self::default().size_of_parent()
+            }
+
             fn delete(mut wid: Self) {
                 assert!(!wid.was_deleted());
                 unsafe {